Move containersource oidc tests
